Split App.tsx into multiple components
