+++ client/src/App.css
@@ -7,7 +7,7 @@ html {
   --color-accent-light: #f8f7f1;
 }
 
-body {
+#root {
   background-color: var(--color-background);
   display: flex;
   flex-direction: column;
@@ -25,104 +25,10 @@ main {
   flex: 1;
 }
 
-footer {
-  background-color: var(--color-accent-dark);
-  box-shadow: 0rem -0.25rem var(--color-primary);
-  width: 100%;
-  padding: 0.5em 0;
-  text-align: center;
-  font-size: 0.75em;
-}
-
-footer>p {
-  opacity: 0.6;
-  margin: 0.2em auto;
-}
-
-footer p,
-footer a {
-  color: var(--color-accent-light);
-}
-
 .unauthenticated-app {
   text-align: center;
 }
 
-header {
-  justify-content: space-between;
-  margin-top: 1em;
-  text-align: center;
-}
-
-header>h1 {
-  margin: 0;
-  color: var(--color-primary);
-  font-weight: normal;
-}
-
-header>p {
-  margin-top: 0.2em;
-}
-
-body>nav {
-  width: 100%;
-  height: 3.25em;
-  padding: 0.5em;
-  box-sizing: border-box;
-  background-color: var(--color-accent-dark);
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-body>nav>span {
-  margin: 0.5em;
-  color: var(--color-accent-light)
-}
-
-body>nav>div,
-body>nav>div img {
-  display: block;
-  height: 100%;
-  border-radius: 50%;
-}
-
-body>nav>div.dropdown-selector {
-  position: relative;
-  display: inline-block;
-}
-
-body>nav .dropdown-content {
-  display: none;
-  position: absolute;
-  right: 0px;
-  min-width: 8em;
-  background: var(--color-accent-light);
-  box-shadow: -0.2em 0.2em 10px 0px rgba(0, 0, 0, 0.2);
-}
-
-body>nav .dropdown-selector:hover .dropdown-content {
-  display: block;
-}
-
-body>nav .dropdown-content p {
-  margin: 0;
-  padding: 1em 0.5em;
-  color: black;
-}
-
-body>nav .dropdown-content a {
-  text-decoration: none;
-}
-
-.sign-in-button {
-  display: block;
-  padding: 0.6em;
-  color: var(--color-background);
-  background-color: var(--color-primary);
-  border-radius: 5px;
-  text-decoration: none;
-}
 
 @media only screen and (min-device-width: 667px) {
   header {
