+++ client/src/components/AppFooter/AppFooter.css
@@ -1,0 +1,18 @@
+footer {
+  background-color: var(--color-accent-dark);
+  box-shadow: 0rem -0.25rem var(--color-primary);
+  width: 100%;
+  padding: 0.5em 0;
+  text-align: center;
+  font-size: 0.75em;
+}
+
+footer>p {
+  opacity: 0.6;
+  margin: 0.2em auto;
+}
+
+footer p,
+footer a {
+  color: var(--color-accent-light);
+}
