+++ client/src/components/AppFooter/AppFooter.tsx
@@ -1,0 +1,12 @@
+import "./AppFooter.css"
+
+function AppFooter() {
+  return (
+    <footer>
+      <p>&copy; <a href="https://twitter.com/mleone5244">Mario Leone</a></p>
+      <p>Money icon by <a href="https://icons8.com">Icons8</a></p>
+    </footer>
+  )
+}
+
+export default AppFooter;
