+++ client/src/components/AppHeader/AppHeader.css
@@ -1,0 +1,15 @@
+header {
+  justify-content: space-between;
+  margin-top: 1em;
+  text-align: center;
+}
+
+header>h1 {
+  margin: 0;
+  color: var(--color-primary);
+  font-weight: normal;
+}
+
+header>p {
+  margin-top: 0.2em;
+}
