+++ client/src/components/AppHeader/AppHeader.tsx
@@ -1,0 +1,12 @@
+import "./AppHeader.css"
+
+function AppHeader() {
+  return (
+    <header className="bound-width">
+      <h1>Expense System</h1>
+      <p>Reimbursement simplified.</p>
+    </header>
+  )
+}
+
+export default AppHeader;
