+++ client/src/components/ProfileBar/ProfileBar.css
@@ -1,0 +1,50 @@
+nav {
+  width: 100%;
+  height: 3.25em;
+  padding: 0.5em;
+  box-sizing: border-box;
+  background-color: var(--color-accent-dark);
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+}
+
+nav>span {
+  margin: 0.5em;
+  color: var(--color-accent-light)
+}
+
+nav>div,
+nav>div img {
+  display: block;
+  height: 100%;
+  border-radius: 50%;
+}
+
+nav>div.dropdown-selector {
+  position: relative;
+  display: inline-block;
+}
+
+nav .dropdown-content {
+  display: none;
+  position: absolute;
+  right: 0px;
+  min-width: 8em;
+  background: var(--color-accent-light);
+  box-shadow: -0.2em 0.2em 10px 0px rgba(0, 0, 0, 0.2);
+}
+
+nav .dropdown-selector:hover .dropdown-content {
+  display: block;
+}
+
+nav .dropdown-content p {
+  margin: 0;
+  padding: 1em 0.5em;
+  color: black;
+}
+
+nav .dropdown-content a {
+  text-decoration: none;
+}
