+++ client/src/components/index.js
@@ -1,0 +1,11 @@
+import ProfileBar from "./ProfileBar/ProfileBar";
+import SignInButton from "./SignInButton/SignInButton";
+import AppFooter from "./AppFooter/AppFooter"
+import AppHeader from "./AppHeader/AppHeader"
+
+export {
+  ProfileBar,
+  SignInButton,
+  AppFooter,
+  AppHeader
+}
