+++ client/src/hooks/index.js
@@ -1,0 +1,6 @@
+import { AuthProvider, useAuth } from "./AuthContext/AuthContext";
+
+export {
+  AuthProvider,
+  useAuth
+}
