+++ client/src/index.js
@@ -11,5 +11,5 @@ ReactDOM.render(
       <App />
     </BrowserRouter>
   </React.StrictMode>,
-  document.body
+  document.getElementById("root")
 );
