+++ client/src/views/AuthCallback/AuthCallback.tsx
@@ -1,0 +1,27 @@
+import { Navigate } from "react-router";
+import { useEffect } from "react";
+import { useAuth } from "hooks";
+
+function AuthCallback() {
+  const auth = useAuth();
+  const signIn = auth.signIn;
+  const code = new URLSearchParams(window.location.search).get("code");
+
+  useEffect(() => {
+    fetch(`/api/token?code=${code}`, {
+      credentials: "include"
+    }).then(response => {
+      if (response.ok) {
+        signIn();
+      }
+    })
+  }, [code, signIn])
+
+  if (auth.getIsSignedIn()) {
+    return <Navigate to="/" />
+  } else {
+    return <p>Signing you in...</p>
+  }
+}
+
+export default AuthCallback;
