+++ client/src/views/index.js
@@ -1,0 +1,5 @@
+import AuthCallback from "./AuthCallback/AuthCallback";
+
+export {
+  AuthCallback
+}
